:wrench: Move Settings struct into separate file and use it in DataManager ctor
