+++ src/Plugin.hpp
@@ -11,6 +11,7 @@
 #include "Display/DisplayNode.hpp"
 #include "Enums.hpp"
 #include "Render/Renderer.hpp"
+#include "Settings.hpp"
 
 #include "../../../src/cs-core/PluginBase.hpp"
 #include "../../../src/cs-utils/DefaultProperty.hpp"
@@ -70,64 +71,6 @@ inline constexpr int SETTINGS_COUNT<DepthMode> = 1;
 
 class Plugin : public cs::core::PluginBase {
  public:
-  struct Settings {
-    struct Data {
-      cs::utils::Property<std::string>        mPath;
-      cs::utils::Property<std::string>        mNamePattern;
-      cs::utils::Property<VolumeFileType>     mType;
-      cs::utils::Property<VolumeStructure>    mStructure;
-      cs::utils::Property<VolumeShape>        mShape;
-      cs::utils::DefaultProperty<std::string> mActiveScalar{""};
-    } mData;
-
-    struct Rendering {
-      cs::utils::DefaultProperty<bool>        mRequestImages{true};
-      cs::utils::DefaultProperty<int>         mResolution{256};
-      cs::utils::DefaultProperty<float>       mSamplingRate{0.05f};
-      cs::utils::DefaultProperty<int>         mMaxPasses{10};
-      cs::utils::DefaultProperty<float>       mDensityScale{1.f};
-      cs::utils::DefaultProperty<bool>        mDenoiseColor{true};
-      cs::utils::DefaultProperty<bool>        mDenoiseDepth{true};
-      cs::utils::DefaultProperty<DepthMode>   mDepthMode{DepthMode::eNone};
-      cs::utils::DefaultProperty<std::string> mTransferFunction{"BlackBody.json"};
-    } mRendering;
-
-    struct Lighting {
-      cs::utils::DefaultProperty<bool>  mEnabled{false};
-      cs::utils::DefaultProperty<float> mSunStrength{1.f};
-      cs::utils::DefaultProperty<float> mAmbientStrength{.5f};
-    } mLighting;
-
-    struct Display {
-      cs::utils::DefaultProperty<bool>        mPredictiveRendering{false};
-      cs::utils::DefaultProperty<bool>        mReuseImages{false};
-      cs::utils::DefaultProperty<bool>        mDepthData{true};
-      cs::utils::DefaultProperty<bool>        mDrawDepth{false};
-      cs::utils::DefaultProperty<DisplayMode> mDisplayMode{DisplayMode::eMesh};
-    } mDisplay;
-
-    struct Transform {
-      cs::utils::Property<std::string>       mAnchor;
-      cs::utils::DefaultProperty<glm::dvec3> mPosition{glm::dvec3(0, 0, 0)};
-      cs::utils::DefaultProperty<double>     mScale{1.};
-      cs::utils::DefaultProperty<glm::dvec3> mRotation{glm::dvec3(0, 0, 0)};
-    } mTransform;
-
-    struct Core {
-      cs::utils::DefaultProperty<bool>        mEnabled{true};
-      cs::utils::DefaultProperty<std::string> mScalar{""};
-      cs::utils::Property<float>              mRadius;
-    };
-    std::optional<Core> mCore;
-
-    struct Pathlines {
-      cs::utils::Property<std::string>  mPath;
-      cs::utils::DefaultProperty<bool>  mEnabled{true};
-      cs::utils::DefaultProperty<float> mLineSize{1.f};
-    };
-    std::optional<Pathlines> mPathlines;
-  };
-
   void init() override;
   void deInit() override;
   void update() override;
@@ -163,7 +106,7 @@ class Plugin : public cs::core::PluginBase {
 
     std::string_view      mName;    /// Name of the parameter's UI callback
     std::string_view      mComment; /// Description of the parameter
-    std::optional<Target> mTarget;  /// Property in the Plugin::Settings struct
+    std::optional<Target> mTarget;  /// Property in the Settings struct
     std::optional<Setter> mSetter;  /// Setter method on the Renderer class
     std::optional<Action> mAction;  /// Setter method on the Plugin class
   };
